server/requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
livereload==2.5.1
MarkupSafe==1.0
six==1.10.0
tornado==4.5.1
Werkzeug==0.12.2
python-dateutil==2.6.1
